Do task

+++ src/main/java/ru/kata/academy/kovtunenko/second/block/repository/UserRepositoryImpl.java
@@ -3,20 +3,21 @@ package ru.kata.academy.kovtunenko.second.block.repository;
 import org.slf4j.Logger;
 
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.TransactionException;
 import ru.kata.academy.kovtunenko.second.block.model.User;
 
 import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public class UserRepositoryNoJpa {
-    private static final Logger LOGGER = LoggerFactory.getLogger(UserRepositoryNoJpa.class);
-    @Autowired
+public class UserRepositoryImpl implements UserRepository {
+    private static final Logger LOGGER = LoggerFactory.getLogger(UserRepositoryImpl.class);
+    @PersistenceContext
     EntityManager em;
 
     public List<User> findAll() {
@@ -40,7 +41,7 @@ public class UserRepositoryNoJpa {
     public void save(User user) {
         try {
             em.persist(user);
-        } catch (IllegalStateException | PersistenceException | IllegalArgumentException ex) {
+        } catch (IllegalStateException | PersistenceException | IllegalArgumentException | TransactionException ex) {
             LOGGER.error("Exception when save user", ex);
         }
     }
@@ -48,16 +49,19 @@ public class UserRepositoryNoJpa {
     public void update(User user) {
         try {
             em.merge(user);
-        } catch (IllegalStateException | PersistenceException | IllegalArgumentException ex) {
+        } catch (IllegalStateException | PersistenceException | IllegalArgumentException | TransactionException ex) {
             LOGGER.error("Exception when update user", ex);
         }
     }
 
-
     public void delete(User user) {
         try {
-            em.remove(user);
-        } catch (IllegalStateException | PersistenceException | IllegalArgumentException ex) {
+            if(user.getClass().equals(User.class)) {
+                em.remove(user);
+            } else {
+                throw new IllegalArgumentException("The user variable is not an instance of the User class");
+            }
+        } catch (IllegalStateException | PersistenceException | IllegalArgumentException | TransactionException ex) {
             LOGGER.error("Exception when delete user", ex);
         }
     }
